requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
pydantic-settings==2.1.0
alpaca-py==0.11.0
asyncpg==0.29.0
pandas==2.0.3
numpy==1.24.4
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.8.6
beautifulsoup4==4.12.3
lxml==5.3.0
certifi==2024.8.30
numexpr==2.8.6
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1
